upload cải thiện code

+++ Admin/Category/categories.php
@@ -1,4 +1,8 @@
 <?php
+  session_start();
+  if(!isset($_SESSION['isLogin']) && $_SESSION['isLogin']!= true){
+      header('Location: ../Login/login.php');
+  }
 	//Liên kết tới file Connection
 	require_once('../../connection.php');
 	//lấy danh mục
@@ -27,17 +31,13 @@
 
     <!-- Latest compiled and minified JavaScript -->
     <script src="http://netdna.bootstrapcdn.com/bootstrap/3.1.1/js/bootstrap.min.js"></script>
-    <script language="javascript" text="text/javascript">
-        <form id="vidu2">
-            <input type="button" id="Doimatkhau" value="Đổi mật khẩu" onclick="traloi=confirm('Bạn có muốn đổi mật khẩu ');"/>
-        </form>
-    </script>
 </head>
 <body>
     <div class="container">
     <h3 align="center">Zent - Education And Technology Group</h3>
     <h3 align="center">Category List</h3>
     <a href="category_add.php" type="button" class="btn btn-primary">Thêm mới</a>
+    <a href="../Login/index.php" type="button" class="btn btn-primary">Về trang Admin</a>
     <?php if(isset($_COOKIE['msg'])) { ?>
             <div class="alert alert-success">
                 <strong>Thông báo</strong> <?=$_COOKIE['msg'] ?>
@@ -62,7 +62,7 @@
           <td>
             <a href="category_detail.php?id=<?=$status['id']?>" type="button" class="btn btn-default">Xem</a>
             <a href="category_edit.php?id=<?=$status['id']?>" type="button" class="btn btn-success">Sửa</a>
-            <a href="category_delete.php?id=<?=$status['id']?>" type="button" class="btn btn-warning">Xóa</a>
+            <a href="category_delete.php?id=<?=$status['id']?>"  onclick="return confirm('Bạn có thật sự muốn xóa ?');" type="button" class="btn btn-warning">Xóa</a>
           </td>
         </tr>
         <?php } ?>
